Refactor: reorganiza estrutura dos componentes de residências do frontend
