+++ frontend/src/app/pages/residences/residence-details/residence-details.component.html
