+++ frontend/src/app/pages/residences/residence-list/residence-list.component.html
