+++ frontend/src/app/pages/students/student-list/student-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, inject } from '@angular/core';
 import { StudentService } from '../../../services/student.service';
-import { Student } from '../../../models/student.model';
+import { Student } from '../../../model/student.model';
 import { CommonModule } from '@angular/common';
 import { Router } from '@angular/router';
