+++ frontend/src/app/services/student.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Student } from '../models/student.model';
+import { Student } from '../model/student.model';
 
 @Injectable({
   providedIn: 'root'
